Make the data processing part of the visualisation
This moves Redmine.js to be a fairly generic way to make RESTful calls
to Redmine, which makes a lot more sense IMO. Separate out the
visualisation class into initialisation, data loading (which calls
Redmine and the visualisation method, which draws the data). Data
processing currently happens somewhere in between load() and
visualise().